+++ packages/lumx-core/src/js/utils/locale/getCurrentLocale.test.ts
@@ -1,4 +1,5 @@
 import { vi } from 'vitest';
+
 import { getCurrentLocale } from './getCurrentLocale';
 
 describe('getCurrentLocale', () => {