+++ packages/lumx-react/src/components/time-picker-field/index.ts
@@ -1,0 +1,5 @@
+export { TimePickerField, type TimePickerFieldProps } from './TimePickerField';
+export {
+    COMPONENT_NAME as TIME_PICKER_FIELD_COMPONENT_NAME,
+    CLASSNAME as TIME_PICKER_FIELD_CLASSNAME,
+} from '@lumx/core/js/components/TimePickerField';